--- gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Mon Aug 03 14:48:45 CEST 2020
-distributionUrl=https\://services.gradle.org/distributions/gradle-6.1.1-all.zip
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStorePath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
