+++ src/main/kotlin/de/twintorx/battleship/Main.kt
@@ -1,6 +1,6 @@
 package de.twintorx.battleship
 
-import de.twintorx.battleship.game.Player
+import de.twintorx.battleship.ui.Player
 import org.fusesource.jansi.AnsiConsole
 
 object Main {
