+++ src/main/kotlin/de/twintorx/battleship/game/board/GameBoard.kt
@@ -1,9 +1,7 @@
 package de.twintorx.battleship.game.board
 
-import de.twintorx.battleship.game.Cell
-import de.twintorx.battleship.game.Move
-import de.twintorx.battleship.game.Ship
-import de.twintorx.battleship.game.ShipType
+import de.twintorx.battleship.game.ship.Ship
+import de.twintorx.battleship.game.ship.ShipType
 import java.awt.Point
 
 class GameBoard(
