+++ src/main/kotlin/de/twintorx/battleship/game/board/Move.kt
@@ -1,0 +1,9 @@
+package de.twintorx.battleship.game.board
+
+enum class Move {
+    INVALID,
+    HIT,
+    SUNK,
+    NO_HIT,
+    GAME_OVER
+}
