+++ src/test/kotlin/de/twintorx/battleship/GameTest.kt
@@ -1,11 +1,11 @@
 package de.twintorx.battleship
 
-import de.twintorx.battleship.game.Cell
-import de.twintorx.battleship.game.Move
-import de.twintorx.battleship.game.Ship
+import de.twintorx.battleship.game.board.Cell
+import de.twintorx.battleship.game.board.Move
+import de.twintorx.battleship.game.ship.Ship
 import de.twintorx.battleship.game.board.GameBoard
 import de.twintorx.battleship.game.board.TrackBoard
-import de.twintorx.battleship.game.ShipType
+import de.twintorx.battleship.game.ship.ShipType
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
 import java.awt.Point
